update shopping cart stock

+++ cars/context_processor.py
@@ -5,8 +5,8 @@ def total_cart_value(request):
     total_value = 0
 
     for key, value in cart.items():
-        price = float(value.get("price", 0))
-        quantity = float(value.get("quantity", 0))
+        price = int(value.get("price", 0))
+        quantity = int(value.get("quantity", 0))
         total_value += price + quantity
 
     return {'total_cart_value': total_value}
